# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Laravel Books",
        "description": "Documentation complète de l'API de gestion de livres avec authentification par token (Laravel Sanctum)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Serveur de développement local"
        }
    ],
    "paths": {
        "/books": {
            "get": {
                "tags": [
                    "Livres"
                ],
                "summary": "Liste tous les livres",
                "description": "Retourne une liste paginée de tous les livres (2 par page)",
                "operationId": "f31a4a8c234eea8c0b2fdcd6634c37ce",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de la page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des livres récupérée avec succès"
                    }
                }
            },
            "post": {
                "tags": [
                    "Livres"
                ],
                "summary": "Créer un nouveau livre",
                "description": "Ajoute un nouveau livre (authentification requise)",
                "operationId": "3f946db73e3a02e8e2f56f60c398986b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "author",
                                    "summary",
                                    "isbn"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "minLength": 3,
                                        "example": "1984"
                                    },
                                    "author": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "minLength": 3,
                                        "example": "George Orwell"
                                    },
                                    "summary": {
                                        "type": "string",
                                        "maxLength": 500,
                                        "minLength": 10,
                                        "example": "Un roman dystopique..."
                                    },
                                    "isbn": {
                                        "type": "string",
                                        "maxLength": 13,
                                        "minLength": 13,
                                        "example": "9780451524935"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Livre créé avec succès"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/books/{id}": {
            "get": {
                "tags": [
                    "Livres"
                ],
                "summary": "Afficher un livre spécifique",
                "description": "Retourne les détails d'un livre (avec cache de 60 minutes)",
                "operationId": "3184b7ae54e17e29fd5fa3027972790e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du livre",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Livre récupéré avec succès"
                    },
                    "404": {
                        "description": "Livre non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Livres"
                ],
                "summary": "Mettre à jour un livre",
                "description": "Modifie un livre existant (authentification requise)",
                "operationId": "63f8f46953b2690763bfbaa581da0668",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du livre",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Nouveau titre"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "Nouvel auteur"
                                    },
                                    "summary": {
                                        "type": "string",
                                        "example": "Nouveau résumé..."
                                    },
                                    "isbn": {
                                        "type": "string",
                                        "example": "9781234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Livre mis à jour"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Livre non trouvé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Livres"
                ],
                "summary": "Supprimer un livre",
                "description": "Supprime un livre (authentification requise)",
                "operationId": "c286320eba3b96b72b7baf6cd634ac1f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du livre",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Livre supprimé avec succès"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Livre non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Livres"
                ],
                "summary": "Mettre à jour partiellement un livre",
                "description": "Modifie partiellement un livre (authentification requise)",
                "operationId": "ce23b97b4133915dbb90348f067a375d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du livre",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Nouveau titre"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "Nouvel auteur"
                                    },
                                    "summary": {
                                        "type": "string",
                                        "example": "Nouveau résumé..."
                                    },
                                    "isbn": {
                                        "type": "string",
                                        "example": "9781234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Livre mis à jour"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Livre non trouvé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entrez votre token d'authentification au format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}